refactor(contracts): host expandRunsForInlineNewlines (SD-2836)
Move the inline-newline run expander from @superdoc/pm-adapter into
@superdoc/contracts. The function is a pure transformation on Run[]
shapes already defined here; relocating it lets painter-dom consume the
helper without depending on pm-adapter (per the SD-2836 acceptance
criterion: no painter-dom imports from pm-adapter or layout-bridge).

pm-adapter chains its public re-export through contracts, keeping the
import path stable until painter-dom is migrated to consume directly.
