+++ packages/layout-engine/contracts/src/index.ts
@@ -2065,4 +2065,8 @@ export type {
 } from './resolved-layout.js';
 export { isResolvedTableItem, isResolvedImageItem, isResolvedDrawingItem } from './resolved-layout.js';
 
+// Pure transformations on inline-run shapes (used by pm-adapter, layout-bridge,
+// and painter-dom). Located in contracts to avoid reverse stage dependencies.
+export { expandRunsForInlineNewlines } from './run-helpers.js';
+
 export * as Engines from './engines/index.js';
