+++ packages/layout-engine/pm-adapter/src/converters/paragraph.ts
@@ -18,6 +18,7 @@ import type {
   TrackedChangeMeta,
   SourceAnchor,
 } from '@superdoc/contracts';
+import { expandRunsForInlineNewlines } from '@superdoc/contracts';
 import type {
   PMNode,
   PMMark,
@@ -219,43 +220,6 @@ export function mergeAdjacentRuns(runs: Run[]): Run[] {
   // Push the last run
   merged.push(current);
   return merged;
-}
-
-/**
- * Expands text runs that contain inline newlines into multiple runs.
- *
- * @param {Run[]} runs - The runs to expand
- * @returns {Run[]} The expanded runs
- */
-export function expandRunsForInlineNewlines(runs: Run[]): Run[] {
-  const result: Run[] = [];
-  for (const run of runs) {
-    const textRun = run as TextRun;
-    if ('text' in run && typeof textRun.text === 'string' && textRun.text.includes('\n')) {
-      const segments = textRun.text.split('\n');
-      let cursor = textRun.pmStart ?? 0;
-      segments.forEach((segment, idx) => {
-        if (segment.length > 0) {
-          result.push({ ...textRun, text: segment, pmStart: cursor, pmEnd: cursor + segment.length });
-          cursor += segment.length;
-        }
-        if (idx !== segments.length - 1) {
-          result.push({
-            kind: 'break',
-            breakType: 'line',
-            pmStart: cursor,
-            pmEnd: cursor + 1,
-            sdt: textRun.sdt,
-            trackedChange: textRun.trackedChange,
-          });
-          cursor += 1;
-        }
-      });
-    } else {
-      result.push(run);
-    }
-  }
-  return result;
 }
 
 /**
