+++ packages/layout-engine/pm-adapter/src/index.ts
@@ -41,7 +41,8 @@ export { SectionType } from './types.js';
 export { toFlowBlocks, toFlowBlocksMap } from './internal.js';
 
 // Re-export run type guards and run utilities
-export { isTextRun, expandRunsForInlineNewlines } from './converters/paragraph.js';
+export { isTextRun } from './converters/paragraph.js';
+export { expandRunsForInlineNewlines } from '@superdoc/contracts';
 
 // Re-export cache for incremental conversion
 export { FlowBlockCache } from './cache.js';
